+++ libs/components/forms/src/lib/modules/input-box/input-box-adapter.service.ts
@@ -1,4 +1,9 @@
-import { ElementRef, Injectable } from '@angular/core';
+import {
+  ElementRef,
+  Injectable,
+  RendererFactory2,
+  inject,
+} from '@angular/core';
 
 const ARIA_DESCRIBEDBY_ATTR = 'aria-describedby';
 
@@ -7,6 +12,8 @@ const ARIA_DESCRIBEDBY_ATTR = 'aria-describedby';
  */
 @Injectable()
 export class SkyInputBoxAdapterService {
+  #renderer = inject(RendererFactory2).createRenderer(undefined, null);
+
   public focusControl(elRef: ElementRef): void {
     const control = elRef.nativeElement.querySelector('.sky-form-control');
     /* istanbul ignore else */
@@ -63,9 +70,13 @@ export class SkyInputBoxAdapterService {
 
   #setDescribedByIds(inputEl: HTMLElement, describedByIds: string[]): void {
     if (describedByIds.length === 0) {
-      inputEl.removeAttribute(ARIA_DESCRIBEDBY_ATTR);
+      this.#renderer.removeAttribute(inputEl, ARIA_DESCRIBEDBY_ATTR);
     } else {
-      inputEl.setAttribute(ARIA_DESCRIBEDBY_ATTR, describedByIds.join(' '));
+      this.#renderer.setAttribute(
+        inputEl,
+        ARIA_DESCRIBEDBY_ATTR,
+        describedByIds.join(' '),
+      );
     }
   }
 }