+++ libs/components/forms/testing/src/input-box/input-box-harness.ts
@@ -72,7 +72,7 @@ export class SkyInputBoxHarness extends SkyComponentHarness {
     const errors = await this.locatorForAll(
       new HarnessPredicate(SkyStatusIndicatorHarness, {
         selector:
-          'sky-status-indicator:not(sky-input-box-errors sky-status-indicator)',
+          'sky-status-indicator:not(sky-form-error sky-status-indicator)',
       }),
     )();